apps/api/requirements.txt
# Perplex Edge (Lucrix) API - Python dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0
alembic>=1.13.0

# Auth & Security
bcrypt>=4.0.1
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# API Clients & Utilities
httpx>=0.25.2
requests>=2.31.0
websockets>=12.0
python-dateutil>=2.8.2
email-validator>=2.1.0

# Scheduling & Async
apscheduler>=3.10.4
celery>=5.3.6
redis>=5.0.0
celery[redis]>=5.3.6

# Payments
stripe>=7.0.0

# AI / LLM
openai>=1.3.0
groq>=0.4.0

# Integration Extras
pywebpush>=1.14.0
resend>=0.6.0
numpy>=1.26.0
supabase>=2.3.0; python_version < "3.13"

gunicorn==22.0.0

sqlmodel>=0.0.14
statsbombpy>=1.13.0
